Extracted k8s config related code to k8s-config.js
